tutorials/10-finance-portfolio-analyzer-agent/requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
pytest>=7.4.0
